notmuch: refactor Account/Mailbox/Email private data
